Fixed quaternion. Implemented yaw pitch roll, added ypr message.

+++ flight_code/quad/src/MPU6050.h
@@ -904,7 +904,7 @@ class MPU6050 {
             uint8_t dmpGetEIS(int32_t *data, const uint8_t* packet=0);
             
             uint8_t dmpGetEuler(float *data, Quaternion *q);
-            uint8_t dmpGetYawPitchRoll(float *data, Quaternion *q, VectorFloat *gravity);
+            uint8_t dmpGetYawPitchRoll(double *data, Quaternion *q, VectorFloat *gravity);
 
             // Get Floating Point data from FIFO
             uint8_t dmpGetAccelFloat(float *data, const uint8_t* packet=0);
